requirements.txt

langchain_openai>=0.0.1
langchain>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
serpapi>=2.0.0
